# apps/api/src/modules/interviews/dto/update-interview-feedback.dto.ts
import { createZodDto } from "nestjs-zod";
import { updateInterviewFeedbackSchema } from "@aurahire/shared";

export class UpdateInterviewFeedbackDto extends createZodDto(updateInterviewFeedbackSchema) {}